moved cache collctor to jit
